tests/wpt/web-platform-tests/html/browsers/the-window-object/the-windowproxy-object/test-window-proxy-personalbar-manual.html

<!doctype html>
<html lang="en">
<head>
    <meta charset="utf-8" />
    <title>Window Proxy personalbar visible flag Test</title>
    <link rel="author" title="vanessa" href="mailto:vanessaohsy@gmail.com">
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
</head>
<body>
    <h1>Description</h1>
    <p>WindowProxy.personalbar Test</p>

    <h1>Manual Test Steps:</h1>
    <ol>
        <li>Make the personalbar visible in the user agent before executing this test.</li>
        <li>You may need to manually reload afterwards.</li>
    </ol>

    <div id="log"></div>

    <script type="text/javascript" >
    test(function () {
        assert_not_equals(window.personalbar, undefined, "window.personalbar is undefined");
        assert_true(window.personalbar.visible, "window.personalbar.visible");
    });
    </script>
</body>
</html>
